Create README.md
The app is intended to help improve the reading skills of school children. She does this by asking for basic vocabulary. Right at the beginning you can choose in the settings whether the entire basic vocabulary should be queried, or just the words of one or more letters of the alphabet.

The app randomly asks for words from the selected area of ​​the basic vocabulary. The query runs in such a way that a word is only displayed for a short time and the child then has to repeat this word loudly and understandably. The iPad's speech recognition system compares the spoken word with the displayed word.

If the word was not recognized by the system the third time, you have three more options to press the correct button. After the main game round, in another two-minute game round, the unrecognized words can be recognized correctly and additional points can be collected.

Users can also be created in the settings in order to be able to assign the results later. The results can then be emailed to the teacher.

Whole sentences can also be practiced in a further step in the settings.

According to the findings of school pedagogy, the direct and correct recognition of words in the basic vocabulary and complete sentences helps to improve reading ability and reading fluency. Regular practice of this method helps in everyday school life.